Used ajax to retrieve image data and updated README.md with user stories.
