+++ index.html
@@ -39,6 +39,7 @@
     </section>
 
     <script src="scripts/jquery-2.1.4.min.js"></script>
+    <script src="scripts/images.js"></script>
     <script src='scripts/Chart.js'></script>
     <script src='scripts/app.js'></script>
