+++ scripts/app.js
@@ -10,22 +10,6 @@ var ImgObj = function (pImgFileLoc){
 };
 
 //gobal vars
-var imgFileLocations = [
-  "img/658668main_pia15685-full_full.jpg",
-  "img/658682main_pia15686_full.jpg",
-  "img/667877main_Grotzinger-1PIA15690_800-600.jpg",
-  "img/675225main_pia16029-43_800-600.jpg",
-  "img/676026main_pia16052-color-43_800-600.jpg",
-  "img/676039main_pia16053-color-43_800-600.jpg",
-  "img/678276main_pia16077-43_800-600.jpg",
-  "img/734386main_pia16768-43_800-600.jpg",
-  "img/PIA17766-800x600.jpg",
-  "img/PIA17946-800x600.jpg",
-  "img/PIA18083-800x600.jpg",
-  "img/PIA19803-800x600.jpg",
-  "img/PSP_008579_9020_descent_800-600.jpg",
-  "img/glacialcraters_mro.jpg"
-];
 var didReset = true;
 var whichSitesNow = [0 , 1];
 var imgArray = [];
@@ -51,6 +35,13 @@ var VoteTracker = function() {
   };
   this.displayImg();
 };
+
+VoteTracker.prototype.getImages = function() {
+  $.ajax( {
+    url: 'scripts/images.js',
+    method: 'GET'
+  });
+}
 
 VoteTracker.prototype.makeChart = function () {
   // var ctx = $('#voteChart').get(0).getContext('2d');
@@ -141,7 +132,7 @@ function handleTheReset (event) {
   $('#clickOne').addClass(null);
   $('#clickTwo').html(null);
   $('#clickTwo').addClass(null);
-  
+
   pageOneTracker.displayImg();
 }
